% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pars_permod.R
\name{calc_pars_permod}
\alias{calc_pars_permod}
\title{Calculates weighted parameters per moderator value}
\usage{
calc_pars_permod(moderators, kernel, bandwidth, parameters)
}
\arguments{
\item{moderators}{moderator values for which parameters have to be estimated}

\item{kernel}{kernel}

\item{bandwidth}{bandwidth}

\item{parameters}{dataframe with estimated parameters for moderators in moderator grid}
}
\value{
dataframe with parameter values per moderator value
}
\description{
Calculates weighted parameters per moderator value
}
\examples{

data(simplefactordata)

lavmodel <- "
        F=~ indicator1 + indicator2 + indicator3 + indicator4
        F ~~ 1*F"

model <- train_lsemmodel( simplefactordata, moderator="moderator",
moderator.grid=c(1:9)/10,
lavmodel=lavmodel, bandwidth =2)

calc_pars_permod(moderators =  unique(simplefactordata$moderator),
bandwidth = 0.2, parameters = model$parameters, kernel = "gaussian" )



}
