% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pars_permod.R
\name{calculate_parameters_permoderator}
\alias{calculate_parameters_permoderator}
\title{Calculates weighted parameters per moderator value}
\usage{
calculate_parameters_permoderator(
  moderators_test_data,
  kernel,
  bandwidth,
  parameters_train_data
)
}
\arguments{
\item{moderators_test_data}{moderator values for which parameters have to be estimated}

\item{kernel}{kernel}

\item{bandwidth}{bandwidth}

\item{parameters_train_data}{dataframe with estimated parameterer in moderator grid}
}
\value{
dataframe with parameter values per moderator value
}
\description{
Calculates weighted parameters per moderator value
}
\examples{

data(simplefactordata)

lavmodel <- "
        F=~ indicator1 + indicator2 + indicator3 + indicator4
        F ~~ 1*F"

model <- train_lsemmodel( simplefactordata, moderator_name="moderator",
moderator_grid=c(1:9)/10,
lavmodel=lavmodel, bandwidth =2)

calculate_parameters_permoderator(moderators_test_data =  unique(simplefactordata$moderator),
bandwidth = 0.2, parameters_train_data = model$parameters, kernel = "gaussian" )



}
