% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RAM.R
\name{calculate_RAM}
\alias{calculate_RAM}
\title{Calculates RAM matrices from dataframe of fittted values}
\usage{
calculate_RAM(df_pars_onemod)
}
\arguments{
\item{df_pars_onemod}{df of estimated parameters of lsem model for one value of the moderator}
}
\value{
list:
value_moderator = value of moderator for the matrices;
vector_means = estimated means of multivariate normal;
matrix_covariance = estimated covariance of multivariate normal;
vector_variablenames = sequence variable names in which all matrices / vectors are ordered;
}
\description{
Calculates RAM matrices from dataframe of fittted values
}
