% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RAM.R
\name{calculate_RAM_matrices}
\alias{calculate_RAM_matrices}
\title{Calculates RAM matrices from dataframe of fittted values}
\usage{
calculate_RAM_matrices(parameters_onemoderator)
}
\arguments{
\item{parameters_onemoderator}{df of estimated parameters of lsem model for one value of the moderator}
}
\value{
list:
value_moderator = value of moderator for the matrices;
expected_means = estimated means of multivariate normal;
matrix_covariance = estimated covariance of multivariate normal;
ordered_variablenames = sequence variable names in which all matrices / vectors are ordered;
}
\description{
Calculates RAM matrices from dataframe of fittted values
}
