% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crossvaldata.R
\name{create_crossvaldata}
\alias{create_crossvaldata}
\title{Creates list of datasets from a dataset}
\usage{
create_crossvaldata(data, K)
}
\arguments{
\item{data}{dataframe}

\item{K}{numeric}
}
\value{
list of training and test sets
}
\description{
Creates list of datasets for K-fold crossvalidaton
labelled as training and test for cross-validation
}
\examples{
create_crossvaldata(rbind(rnorm(10), runif(10)), 2)

}
