% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_model.R
\name{CV_model}
\alias{CV_model}
\title{Cross validates lavaan model over list with bandwidths and training and test sets}
\usage{
CV_model(list, moderator, moderator.grid, lavmodel, kernel, ...)
}
\arguments{
\item{list}{list with bandwidths and training and test sets}

\item{moderator}{name of moderator}

\item{moderator.grid}{grid for moderator values}

\item{lavmodel}{lavaan model}

\item{kernel}{kernel used}

\item{...}{further arguments to be passed to lavaan::sem or lavaan::lavaan}
}
\value{
dataframe of bandwidths and likelihoods
}
\description{
Cross validates lavaan model over list with bandwidths and training and test sets
}
\examples{
data(simplefactordata)

lavmodel <- "
        F=~ indicator1 + indicator2 + indicator3 + indicator4
        F ~~ 1*F"

bandwidthvector <- c(1,2)



list <- purrr::cross2(create_crossvaldata(data = simplefactordata, K = 10),
   bandwidthvector)

CV_model(list = list[[1]], moderator = "moderator",
       moderator.grid =c(1:9)/10,
       lavmodel = lavmodel, kernel ="gaussian")


}
