% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsem_kernel_weights.R
\name{lsem_kernel_weights}
\alias{lsem_kernel_weights}
\title{Function from sirt that calculates kernel weights for certain bandwidth and moderator value}
\usage{
lsem_kernel_weights(x, x0, bw, kernel = "gaussian")
}
\arguments{
\item{x}{vector of datapoints to calculate weights}

\item{x0}{moderator value}

\item{bw}{bandwidth}

\item{kernel}{kernel type, defaults to gaussion}
}
\value{
vector of weights
}
\description{
Function from sirt that calculates kernel weights for certain bandwidth and moderator value
}
