% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_full_matrix.R
\name{make_full_matrix}
\alias{make_full_matrix}
\title{This functions makes a matrix with same sequence of colnames and rownames
from a dataframe of fitted values}
\usage{
make_full_matrix(df_fittedvalues, vector_variablenames)
}
\arguments{
\item{df_fittedvalues}{dataframe of fitted values of one type (regression and loadings or (co-)variance)}

\item{vector_variablenames}{sequence of variable names to ensure compatability of all matrices}
}
\value{
a matrix with same sequence of colnames and rownames as vector_variablenames
and values from the dataframe of fitted values and everything else 0
}
\description{
This functions makes a matrix with same sequence of colnames and rownames
from a dataframe of fitted values
}
