% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Function to simulate modulated simple factor model with 4 indicators}
\usage{
simulate_data(type, effect, N, seed, path = NA, i = NA)
}
\arguments{
\item{type}{type of distribution of the moderator, can be "rnorm", "mixture",
"increasing","unbalanced mixture"}

\item{effect}{whether the effect of the moderator is "equal" between loading 3 and 4
or "ascending"}

\item{N}{nrow of simulated dataset}

\item{seed}{seed for simulation}

\item{path}{where to write the dataset}

\item{i}{extra variabele to give simulated dataset a number}
}
\value{
simulated dataset
}
\description{
Function to simulate modulated simple factor model with 4 indicators
}
\examples{

simulate_data(type = "rnorm", effect = "equal", N = 100, seed = 121212)
}
