% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bandwidths.R
\name{test_bandwidths}
\alias{test_bandwidths}
\title{Fits LSEM models with predifined bandwidths and measures fit}
\usage{
test_bandwidths(
  data,
  lavmodel,
  bandwidthvector,
  moderator,
  moderator.grid,
  statistic = "AIC",
  ...
)
}
\arguments{
\item{data}{dataset for model}

\item{lavmodel}{Specified SEM in lavaan}

\item{bandwidthvector}{vector of tested bandwidths}

\item{moderator}{Variable name of the moderator}

\item{moderator.grid}{Focal points at which the LSEM should be evaluated.}

\item{statistic}{can be AIC or CV (cross-validated likelihood)}

\item{...}{further arguments to be passed to lavaan::sem or lavaan::lavaan.}
}
\value{
dataframe with fit measures for every bandwidht
}
\description{
Fits LSEM models with predifined bandwidths and measures fit
}
\examples{

data(simplefactordata)

lavmodel <- "
        F=~ indicator1 + indicator2 + indicator3 + indicator4
        F ~~ 1*F"

test_bandwidths(simplefactordata,moderator="moderator",moderator.grid=c(1:9)/10,
lavmodel=lavmodel, bandwidthvector = c(1,2), statistic = "AIC")

}
