% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_model.R
\name{CV_model}
\alias{CV_model}
\title{Cross validates lavaan model over list with bandwidths and training and test sets}
\usage{
CV_model(
  list,
  moderator_name,
  moderator_grid,
  lavmodel,
  kernel,
  digits,
  warningabort = FALSE,
  silent = FALSE,
  ...
)
}
\arguments{
\item{list}{list with bandwidths and training and test sets}

\item{moderator_name}{name of moderator}

\item{moderator_grid}{grid for moderator values}

\item{lavmodel}{lavaan model}

\item{kernel}{kernel used}

\item{digits}{number of digits test_data moderator is rounded to, can be lowerd to increase speed}

\item{warningabort}{whether to stop further calculations when lavaan model fitting produces
a warning}

\item{silent}{if false print output in the meantime (false = default)}

\item{...}{further arguments to be passed to lavaan::sem or lavaan::lavaan}
}
\value{
dataframe of bandwidths and likelihoods
}
\description{
Cross validates lavaan model over list with bandwidths and training and test sets
}
\examples{
data(simplefactordata)

lavmodel <- "
        F=~ indicator1 + indicator2 + indicator3 + indicator4
        F ~~ 1*F"

bandwidthvector <- c(1,2)



list <- purrr::cross2(create_crossvaldata(data = simplefactordata, K = 10),
   bandwidthvector)

CV_model(list = list[[1]], moderator_name = "moderator",
       moderator_grid =c(1:9)/10,
       lavmodel = lavmodel, kernel ="gaussian")


}
