% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bandwidths.R
\name{test_bandwidths}
\alias{test_bandwidths}
\title{Fits LSEM models with predifined bandwidths and measures fit}
\usage{
test_bandwidths(
  data,
  lavmodel,
  bandwidthvector,
  moderator_name,
  moderator_grid,
  statistic = "AIC",
  K = 10,
  kernel = "gaussian",
  digits = 6,
  maxbandwidthdistance = 2,
  ...
)
}
\arguments{
\item{data}{dataset for model}

\item{lavmodel}{Specified SEM in lavaan}

\item{bandwidthvector}{vector of tested bandwidths}

\item{moderator_name}{Variable name of the moderator}

\item{moderator_grid}{Focal points at which the LSEM should be evaluated.}

\item{statistic}{can be AIC or CV (cross-validated likelihood)}

\item{K}{number of folds for cross-validation, defaults to 10}

\item{kernel}{used kernel to weight observations}

\item{digits}{number of digits test_data moderator is rounded to,
can be lowered to increase speed crossvalidation
defaults to 6}

\item{maxbandwidthdistance}{the amount of bandwidths from every point in the
moderator grid that a point in the dataset must be present}

\item{...}{further arguments to be passed to lavaan::sem or lavaan::lavaan.}
}
\value{
dataframe with fit measures for every bandwidth
}
\description{
Fits LSEM models with predifined bandwidths and measures fit
}
\examples{

data(simplefactordata)

lavmodel <- "
        F=~ indicator1 + indicator2 + indicator3 + indicator4
        F ~~ 1*F"

test_bandwidths(simplefactordata,moderator_name="moderator",moderator_grid=c(1:9)/10,
lavmodel=lavmodel, bandwidthvector = c(1,2), statistic = "AIC")

}
