% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_lsemmodel.R
\name{train_lsemmodel}
\alias{train_lsemmodel}
\title{Fits LSEM model on training set with predefined bandwidth}
\usage{
train_lsemmodel(data, lavmodel, bandwidth, moderator_name, moderator_grid, ...)
}
\arguments{
\item{data}{training dataset}

\item{lavmodel}{Specified SEM in lavaan}

\item{bandwidth}{Bandwidth factor}

\item{moderator_name}{Variable name of the moderator}

\item{moderator_grid}{Focal points at which the LSEM should be evaluated.}

\item{...}{Further arguments to be passed to lavaan::sem or lavaan::lavaan.}
}
\value{
fitted LSEM model
}
\description{
Fits LSEM model on training set with predefined bandwidth
}
\examples{

data(simplefactordata)

lavmodel <- "
        F=~ indicator1 + indicator2 + indicator3 + indicator4
        F ~~ 1*F"

train_lsemmodel(simplefactordata,moderator_name="moderator",moderator_grid=c(1:9)/10,
lavmodel=lavmodel, bandwidth=2)



}
